Built out most of the page
I built out most of this, and I meant to commit more, however I’ve been
so entranced in this, and it’s been fun. So far I have to get 2 more
images, make this website more responsive, and I have to put in the
content for my resume/portfolio. It is coming along quite well and even
fixed a bug with the background image in Safari.